Update CSS animation to use translateX
This commit adds translateX(150px) to move the main page and open
the menu. translateX is a CSS3 property which can take advantage
of GPU acceleration for smooth animations. Previously the left
and right absolute position properties were being used to move
the menu. These properties cannot be accelerated using the GPU.

translateZ(0) is used on the main-content to prepare the GPU
for the upcoming animation